Widget: Don't check if it's the correct class everywhere.
ELM_WIDGET_DATA_GET was checking it was the correct class. This is
stupid and wrong (and makes tools rightfully complain) because these
functions are implementations of that class anyway, so that would never
be false.
Functions that outside facing (that is, not called from eo_do) should
use API_ENTRY.