conform: Fixed to get a geometry from *focus_region_get for show_region_job.
Summary:
The geometry of elm_widget_show_region_get is only updated by elm_widget_show_region_set.
If the geometry of *show_region_get is not updated, the result of show_region_job would be incorrect.
So, we need to get a proper geometry using *focus_region_get for focused object.

Reviewers: seoz, woohyun, Hermet, raster, tasn

Reviewed By: raster

Differential Revision: https://phab.enlightenment.org/D466